+++ lib/cache/index.dart
@@ -1,0 +1,1 @@
+export 'service/cache_service.dart';