+++ lib/dialog/index.dart
@@ -1,0 +1,1 @@
+export 'service/dialog_service.dart';