+++ lib/image/index.dart
@@ -1,2 +1,4 @@
 export 'img_view.dart';
 export 'photo_galley_screen.dart';
+export 'image_crop_screen.dart';
+export 'service/image_crop_service.dart';