+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0-dev.1
+
+ADD:
+- image crop screen/services.
+- picker services.
+- cache services.
+
 ## 0.1.0-dev.2
 
 fix typos.